+++ zbus/src/object_server/interface/interface_deref.rs
@@ -3,7 +3,7 @@ use std::{
     ops::{Deref, DerefMut},
 };
 
-use crate::async_lock::{RwLockReadGuard, RwLockWriteGuard};
+use crate::runtime::async_lock::{RwLockReadGuard, RwLockWriteGuard};
 
 use super::Interface;
